agregado de imagenes a una cabania + bootstrap

+++ api/ComentariosCabaniaApi.php
@@ -1,16 +1,19 @@
 <?php
 
 require 'api.php';
-require '../model/ComentarioModel.php';
+include_once (dirname(__DIR__).'/model/ComentarioModel.php');
+include_once (dirname(__DIR__).'/controller/AdminController.php');
 
 class ComentarioApi extends Api
 {
   private $model;
+  private $adminController;
 
   public function __construct($request)
  {
     parent::__construct($request);
     $this->model = new ComentarioModel();
+    $this->adminController = new AdminController();
   }
 
   protected function comentarios($argumentos){
@@ -20,21 +23,25 @@ class ComentarioApi extends Api
         break;
 
       case 'DELETE':
+      if ($this->adminController->checkPrivilegio() == 1) {
             if(count($argumentos)>0){
               $error['Error'] = "El comentario no existe";
               $success['Success'] = "El comentario se borro";
               $filasAfectadas = $this->model->eliminarComentario($argumentos[0]);
                 return ($filasAfectadas == 1) ? $success : $error;
               }
+            }
         break;
 
         case 'POST':
-            if((count($argumentos)==0) && ($_POST["admin"])){
+          if ($this->adminController->checkPrivilegio() >= 0) {
+            if(count($argumentos)==0){
               if (isset($_POST["texto"]) && ($_POST["texto"] != "")) {
                 $error['Error'] = "El comentario no se creo";
                 $id_comentario = $this->model->crearComentario($_POST["texto"],$_POST["puntaje"],$_POST["id_cabania"]);
                 }
               }
+            }
               return ($id_comentario > 0) ? $this->model->getComentario($id_comentario) : $error;
 
           break;
